Externalize API connection
Externalization of the API call in the form of a service within the component
