+++ src/public/characters/characters.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { ajax } from 'rxjs/ajax';
+import { CharactersService } from './services/characters.service';
 
 @Component({
   selector: 'app-characters',
@@ -9,21 +9,8 @@ import { ajax } from 'rxjs/ajax';
 export class CharactersComponent implements OnInit {
   page: number = 1;
   characters = new Array<any>();
-  constructor() {
-    for (let page = 1; page < 42 + 1; page++) {
-      const apiData = ajax(
-        `https://rickandmortyapi.com/api/character/?page=${page}`
-      );
-      apiData.subscribe((res: any) => {
-        for (
-          let character = 0;
-          character < res.response['results'].length;
-          character++
-        ) {
-          this.characters.push(res.response['results'][character]);
-        }
-      });
-    }
+  constructor(charactersService: CharactersService) {
+    this.characters = charactersService.getCharacters();
   }
 
   ngOnInit(): void {}
